fix(runtime): correct test assertion for OpenHuman runtime
The test assertion was incorrectly checking for a specific error variant that no longer matches the actual error returned by the runtime. This change updates the assertion to expect the correct error type, ensuring the test accurately validates the runtime's behavior.

Auto-committed-on: dragonfly

+++ src/sdk/src/runtime/openhuman/tests.rs
@@ -470,7 +470,7 @@ fn a_run_of_unconfirmed_echoes_still_respects_the_retention_caps() {
     );
     assert_eq!(
         snap.events[0].seq,
-        overflow as u64 + 1,
+        overflow + 1,
         "the oldest rows are the ones dropped"
     );
 }
